Expand the universe of options attempted for explicit type variables
Summary:
The core idea of solving for explicit type variables is that we try each of a finite number of options against the gathered bounds, and then instantiate it with the first one that works.
However, in theory, the domain of these variables is infinite: they could be instantiated with any of the finite options they're declared with, as well as *any* other explicit type variable which contains a subset of the explicit members.  We were previously approximating this set with the lower bound if that lower bound was a valid explicit variable.

However this wasn't addressing the scenario where the upper bound had the variable in it, let alone if that upper bound was a union of that variable and other things.

To address this, I made a new approximation: all valid explicits mentioned anywhere in either bound.
This should in theory list all of the possible relevant elements in that infinite set.

Reviewed By: dkgi

Differential Revision: D21699206

fbshipit-source-id: f71fa0d648153e5198f45c1bbde809400ff7886e