converting `metaclass` to take a class name
Summary:
It turns out that all of the uses of `metaclass` have to look up the class definition from the name anyway: so it cleans things up to just put that lookup inside of the method.
The only call that doesn't already come immediately after a lookup is the recursive call inside metaclass itself.  This is handled by just making the handler recursive.

Reviewed By: pradeep90

Differential Revision: D21760206

fbshipit-source-id: cc473fe5f42006cf4cb27dc2c9734417fe24f65c