Cached metaclass
Summary:
Now that we've simplified `metaclass` down to a string (+ assumptions) -> type function, it's now very easily cachable.
With the class infra in place, we just need to make a new cache in the environment sequence, then add an overload to the hierarchy to use the cache.
I chose to place in between parse_annotation and uninstantiated_attributes, since metaclass always does a parse, and uninstantiated_attributes will always do a metaclass lookup.

Reviewed By: sinancepel

Differential Revision: D21760534

fbshipit-source-id: 334345eead35603ad73e138407cd841d73846010