Separating WeakenMutableLiterals from attributeResolution
Summary:
This module was already isolated, exposed, and had no dependencies on the rest of attributeResolution.  Thus there's no reason why we can't separate it into a different file towards the end of making the big file smaller.

NOTE: If we go down the mixin route proposed in D21769874, this might make sense to promote directly into a mixin.  For now that improves both the status quo, and moves us towards that option anyway.

Reviewed By: pradeep90

Differential Revision: D21770447

fbshipit-source-id: 8ccde220c7043e551ad6dc356d42128cad9c6f30