Resolve attribute access inside of decorator names
Summary:
Simple global lookup isn't actually enough for all use cases for decorators, we also need to handle attributes of globals.  This is doable with all of the resources already available to us in attribute resolution. In effect this is a simplified version of the attribute lookup in TypeCheck, and accordingly doesn't have all of the doodads that that logic does (placeholder stubs, `__getattr__` etc.), but it's enormously more powerful than the status quo.

due to the legwork I've done with unifying constructor logic with callability, this is also enough to support "class-based" decorators ( https://github.com/facebook/pyre-check/issues/195)

This actually is sufficient for us to support the problematic `click` cases where a decorated function becomes a global with attributes which can decorate other functions.

Reviewed By: grievejia

Differential Revision: D21779103

fbshipit-source-id: 45cb012c088393f1ccdab237e7c9d99bfe7c6243