Use global annotation lookup instead of undecorated_signature for resolving decorators
Summary:
with D21775603, we now have access to global annotations inside of AttributeResolution.

This diff takes advantage of that to:
* stop the use of the "undecorated function environment" completely
* support decorated decorators
* support decorators that are globals other than `def`s

The only major wrinkle here is that we can run into infinite loops where resolving a decorator depends on its own resolution.  We break these loops as we have before: with an assumption.  In this case we need to be pessimistic and assume that something is **not** a valid decorator while resolving it.

Reviewed By: grievejia

Differential Revision: D21777421

fbshipit-source-id: b0cf8449afda1d27977f11afd665cf4cc5883f10