Drop `WildcardExports` table
Summary:
The only remaining user of the `WildcardExports` sharedmem table is auto complete. This diff changes autocomplete such that it looks up module metadata instead of wildcard imports, eliminating the need to keep that table.

If one look at the autocomplete code, it is very obvious that what it really wants is an API like `Module.get_all_exports` (see D21676904) but we didn't have that in the past so it's trying to desperately emulate an equivalent with wildcard exports. Now that we have the right API available, let's just use that and simply the logic.

Reviewed By: dkgi

Differential Revision: D21799702

fbshipit-source-id: 2093d24c5467f53c20846443f9dbeab70280cb07