Use `DependencyKey` in `RawSources` table
Summary:
`ReferenceDependencyKey` is a legacy structure that doesn't leverage the new dependency key registry. The only user of it is the `RawSources` table in `AstEnvironment`, and the usage seems to be relatively easy to convert.

This diff basically switches the dependency key type for `RawSources` from the legacy `ReferenceDependencyKey` to the better-supported `DependencyKey`. The change is mostly massaging the types whereas the basic logic should remain the same.

Reviewed By: dkgi

Differential Revision: D21826123

fbshipit-source-id: e11d5f73ffde9c789e5154e738748d0201bd0589