Resolve globals as arguments to decorator factories
Summary:
Now that we have access to resolving references in the global scope (D21779103), we can start applying that tool in other places.

The first one here is to resolve arguments to decorator factories.  Previously, we only supported pure literal arguments, but with our new tools we can easily support simple names as well.  This, critically, includes enums.

Reviewed By: grievejia

Differential Revision: D21792565

fbshipit-source-id: 7766b9e229ba4231c97e358a696e12415dd9402c