error on non-literal, non-global arguments to decorator factories
Summary:
With the legwork done on D21792565 (*), we can start erroring on the remaining cases.  Without the ability to determine the types of these arguments, we can't reliably determine the correct type for the decorator.  At this point the trade-offs between potentially limited coverage and confusing inconsistency has been tipped towards avoiding inconsistency transparently.

We do so by pre-resolving our arguments, and detecting when one of them is unresolvable, and emitting a corresponding error.

Reviewed By: dkgi

Differential Revision: D21673482

fbshipit-source-id: 5732bbea574d3737f3b82efea5eef3443693bb3e