Route decorator calls as special method __call__ accesses
Summary:
When you call an object in python, the runtime looks up the `__call__` attribute as a "special method" lookup.  This means that if you're accessing through `Type[X]`, they are looked up on the metaclass of `X`, not on the definition of `X`.

This behavior is necessary for class-based callable decorators to work correctly (as the decoration happens with the constructor, not the `__call__` method of the result).

Reviewed By: grievejia

Differential Revision: D21800155

fbshipit-source-id: 82073af0c6e126ad171b93229ef1c4091b4631cd