Coalesce for parameter specs just like for list variadics
Summary: In our quest to make `typing.Callable` just another class, we need to support `[A, B, C]` as a valid inhabitant for a `TParams`.  This diff does just that, as well as supporting the `[A, B, C] -> [[A, B, C]]` conversion that we do for list variadics, but for parameter specifications.

Reviewed By: grievejia

Differential Revision: D21852034

fbshipit-source-id: 56e1a85600a6eca92e4a6cc5fb6d61e8696fcfc3