Include GenericMeta unconditionally
Summary:
PEP 560 removed `typing.GenericMeta`, so it was conditionally removed from typeshed by https://github.com/python/typeshed/pull/3933 .

However, we rely on that class to be the metaclass of `Generic`, all generic classes, and all protocols.  This meant that once typeshed was updated, lots of attributes disappeared.

This diff fixes that by special casing it back in.  This means we'll erroneously allow users to refer to it, but that's fine for now.

Reviewed By: grievejia

Differential Revision: D21854250

fbshipit-source-id: 294740a866800b70939f425e8919a3c275b36285