Error when decorator fails to apply
Summary: This is the very last of the instances I've identified in decorator application where we were silently ignoring failure.  With all of the infrastructure that we built for D21675683, this is a very simple plumbing job of the failure kind, through the attribute/global, through type check to the error display.

Reviewed By: dkgi

Differential Revision: D21676113

fbshipit-source-id: 94e0fd196febaf51a56387e251ef2860cdf823d3