Erroring when factory signature selection fails
Summary:
We previously just silently swallowed signature selection errors from applying a decorator factory.  Now we have all the machinery to surface the errors that we're already producing.

In theory we already have errors for these situations, but just as in D21674228, I think it's better to be safe/explicit than sorry/confusing in the situations where the implementations diverge.

Reviewed By: dkgi

Differential Revision: D21675683

fbshipit-source-id: b641823f01555140a1970f22cc037acbd9c915be