Log recently-used local configuration.
Summary:
Problem: When the user runs a Pyre command from the project root and not a local configuration, we currently just show an error message.

Instead, we can maintain a list of recently-used local configurations. In the above error scenario, we would prompt the user to choose from the list of recent servers.

Reviewed By: dkgi

Differential Revision: D21821547

fbshipit-source-id: cf728cee3cebcc2cca6d2b7bea6d42f8c4837f88