Fix filter directories for targets in configurations
Summary:
Was debugging an unrelated issue when I realized that files I expected to have type checking turned on were not reporting errors.

In the backend these files were marked as "external" - looking closer it turns out that if we have a local directory at `a/.pyre_configuration.local` that contains targets `a/b:c`, our client was creating a filter directory at `a/b/c` rather than `a/b`/.

For more context, when we run `pyre --target //a/b:c` on the command line, we infer the filter directory to be the location of the TARGETS file (`a/b`) because we are given no other information.

However, when a local configuration exists, we should be reporting any errors existing under that local root.

 ---

This bug was introduced because the old logic here used to pass around an `arguments.Namespace` object, so we could simply take a look at `arguments.targets` (aka, targets passed in from the command line) when constructing the filter directory.

Now, `arguments.Namespace` objects are no longer used, so there was confusion between what local `targets` variable represents targets in the command line vs. targets that were already processed and parsed out of the configuration. :(

 ---

Note: the client release that pushes this fix will cause errors to appear in fbcode.

Reviewed By: dkgi

Differential Revision: D21907712

fbshipit-source-id: 0557e7889d7543b7f05e314e692f72de6dbc7635