Remove special case for Type[X] < Callable
Summary:
This is the last use of `constructor` outside of AttributeResolution. Removing the use of constructor when comparing `Type[X] < Callable` and replacing it with the standard `resolve_callable_protocol` does actually change the behavior in situations where `X` has a `__call__` defined on it.  This is good when its on a metaclass (as in D21872355), but problematic when its in the normal hierarchy.

This means that we will problematically prefer the class `__call__` over the metaclass/constructor versions, even though in reality when "special calling" it we'll hit those metaclass versions.
I still think this is the right move because that means we treat literal Callable types and callback protocols consistently, as well as other special methods (e.g. `__getitem__`).

Reviewed By: dkgi

Differential Revision: D21872813

fbshipit-source-id: 3341c947d3ba1625d9e54166f953a26fbe133cf8