Correcting weird meet between Callables and Type[X]
Summary:
I think this case was actually always dead code in prod because we were asking for constructor of the meta type, not the inner type, and constructor was only defined for primitives, which these are not, by definition.

However, it's good that this was broken, because the core idea is wrong: a callable is not less than a meta-type, it's only the other way around.

Thus it makes sense to just delete

Reviewed By: grievejia

Differential Revision: D21872456

fbshipit-source-id: 79274549e6221efa25e9d04710ed6256bbeff255