Handle (m: Type[X]).__call__ for all X
Summary:
I'd like to remove the need to use the `constructor` interface, since in reality it's just a special cased `Type.__call__` implementation that ought to be accessed through the normal attribute interfaces.

A small blocker to this was the fact that the special case I already added there is incomplete, and this diff augments it to handle tuples, and anything else that can be handled by `Type.resolve_class`.

Reviewed By: sinancepel

Differential Revision: D21872268

fbshipit-source-id: ab4e137ad409fb6edcdf9ecdfc32eec4e27d99e8