print the error/file which we error on
Summary:
Right now, the Pyre client doesn't print the entire exception if one is raised (and in our case, the
exception is hard to parse). Let's surface the erroring file + the error message before raising the error (a general fix to also printing uncaught exceptions might also be useful, but likely require a bit more discussion).

Reviewed By: dkgi

Differential Revision: D21943048

fbshipit-source-id: 5c5d6c61eaf200aae0b37a37481ed5cdd5950f81