RFC: Support nominally named parameters preceding parameter spec components
Summary:
While writing up the spec for `Concatenate`, I realized that I have been enforcing a requirement that doesn't actually need to exist.

It is correct that we can't concatenate named parameters onto parameter specifications.  However: this need only apply to the external interface of the function.  As long as we require that the "named" parameters being prepended are being called positionally, the function itself can be written to have named parameters.

Concretely, both
```
   def f(x: X, *args: _TParams.args, **kwargs: _TParams.kwargs) -> R: ...
 ```
and
```
   def g(x: X, \, *args: _TParams.args, **kwargs: _TParams.kwargs) -> R: ...
```
are valid instances of `Callable[Concatenate[X, TParams], R]`

The question is: what is the better user experience:

```
    def q(x: int) -> None: ...

    def f(
        fn: Callable[_TParams, _TReturn], # Option A:  error here and complain that prepended arguments must be positional-only
        *args: _TParams.args,
        **kwargs: _TParams.kwargs,
    ) -> _TReturn: ...

   f(fn=q, x=1) # Option B: Potentially confusingly error here complaining about missing first argument while it seems like you're correctly passing in a fn
```

This diff opts for B, as I feel like it's natural to add named parameters, especially when that named parameter is `self`.  Requiring `__` or `\` seems too painful.

I think we can avoid the confusion in option B with some best-effort special casing on the error to explicitly mention this policy.

Reviewed By: shannonzhu

Differential Revision: D21945652

fbshipit-source-id: 8f215c52f7a4fbbbe821c942e1a88a3e25bdf867