Allow new export resolver to resolve None
Summary:
`None` ought to be mentioned in the typeshed builtins.pyi, but it is not.  Instead, we inject it in unannotatedGlobalEnvironment.  This was sufficiently far back until the introduction of resolve_exports, which operates on module metadata which is upstream of the unannotatedGlobalEnvironment.

The real fix for this is to inject all of our modifications in preprocessing instead of in the unannotatdGlobalEnvironment, but for now its necessary to just special case it into the module metadata.

Reviewed By: grievejia

Differential Revision: D21944733

fbshipit-source-id: c35e32482ce887fc9943ea8c9cce76f7ff0cf33b