eliminate the buck run middleman when generating taint models
Summary:
Now that we're in a situation where we don't ever introspect, the idea of a generated target is unnecessary - we'll rely on Pyre anyway, and the fact that it creates a generated directory to analyze is an implementation detail that doesn't require us to buck run here.

This will also mean that we can turn this into a fbpkg or dotslash it, saving a lot more time when generating fbcode models.

Note: There is a bit of ugliness in the separation between the fbcode_generators and the generate_fbcode_taint_models files - there's no reason, morally speaking, that we can't merge the two. I chose the route of not modifying the fbcode_generators too much now to keep the diff simple to understand, and I feel like there is a little bit of value in keeping the model generator implementations separate from the argparse logic. However, future work can pretty easily refactor these to only keep the dictionary generation in the fbcode_generators file.

This also reveals a flaw in the design of the `run_generators` stuff - we have two main use cases, and the argparse is only shared between the open-source and instagram bits at the moment, so it's probably worth fixing the API up as well.

For now, I'm taking the lazy way out.

Reviewed By: dkgi

Differential Revision: D22027774

fbshipit-source-id: a5b2d64d2e1bbc6efbb8ac96796402ef50a0e5fe