Do not warn on missing annotation for property setter call
Summary: We should not error if a constructor calls the setter of a property attribute of the same name. Now that property attributes take precedent over explicitly assigned attributes in the constructor, `self.x = value` in the constructor is analyzed as a call to the setter of x if it exists and the initialization of a new attribute if x is not a property.

Reviewed By: dkgi

Differential Revision: D21654805

fbshipit-source-id: 910300070320add31706e1370c29abf9a3009aca