Add initial stubs for BaseUserManager and UserManager (#272)
Summary:
These classes should be used whenever the included `User` is extended or
replaced. Like the other stubs for this module, they just establish the
types.

Because `BaseUserManager` belongs in `base_user`, not `models`, the
module it being added and `AbstractBaseUser` is being moved there as
well.
Pull Request resolved: https://github.com/facebook/pyre-check/pull/272

Reviewed By: dkgi

Differential Revision: D22073510

Pulled By: gbleaney

fbshipit-source-id: 4a5f500d1a09a48cc60cf9a81f2eb5aace108203