Print rage for recently-used servers instead of running servers.
Summary:
Problem: We usually want to see the rage for a recently-used server. Earlier, `pyre rage` printed rage only for running servers, which meant that it did nothing for stopped servers, local or root.

This diff prints the rage for the recently-used local roots (or the root server, in case there are no local roots).

Reviewed By: dkgi

Differential Revision: D22098722

fbshipit-source-id: 4e4bc800217d6ffe315beaa3f482fcd61b5432e2