Stop relying on `sexp_of_t` and `t_of_sexp` for `Core.Unix.File_descr.t`
Summary:
`sexp_of_t` and `t_of_sexp` will stop working properly (or even stop to exist) for file descriptors in the next Core upgrade. And this is for good reason -- it is a terrible, terrible idea to persist&serialize file descriptors, because those numbers are not stable at all: If one opens a file/socket, then close it, then open another one a while later, it is highly likely that the file descriptor assigned to the old file will be recycled by the OS and get assigned to the new file/socket as well. The end result is that the descriptor remains the same while the underlying file/connection changes, which could be a huge problem if one assumes that a file descriptor would uniquely identify a file or connection.

This diff makes no attempt at fixing the problem though -- it simply rewrite the existing logic such that it won't break in the upgrade.

Reviewed By: pradeep90

Differential Revision: D22141010

fbshipit-source-id: 6091285620571a9b1b76fef0868ba81ccbb232fe