Enable shared reader mode when acquiring lock.
Summary:
Problem: We would like a reader-writer lock on the shared analysis directory. Readers would be commands like `incremental` and the writer would be the file monitor, which would update files in that directory.

We use the `LOCK_SH` lock command, which acquires the lock as shared reader, as per `fcntl`. [Man page on LOCK_SH](https://linux.die.net/man/2/flock).

Reviewed By: grievejia

Differential Revision: D22149148

fbshipit-source-id: 204cb04c1ea2ac053b37d602b04dd2963fe946c0