Use reader-writer lock for analysis directory and incremental.
Summary:
Problem: `pyre` currently gives stale results when a rebuild is in progress.

We want it to either wait on the rebuild or at least inform the user about the staleness. This diff considers writers to be the file monitor updating the analysis directory (including rebuilds) and readers as incremental queries.

Reviewed By: grievejia

Differential Revision: D22149157

fbshipit-source-id: db7d333784f0945ac360d1d75e0d1c0228b8177b