support sanitizing globals and attributes
Summary:
Context: We've observed a common pattern where request handling code reassigns to a few properties in HttpRequests including GET and POST.
Reassigning to these special attributes of a HttpRequest is a common source of traces becoming artificially longer, usually as a result of a copying or
modification of a user's request data. We avoid simulating these assignments to keep the traces at a shorter length.

This diff introduces the ability to sanitize an attribute or a global, which will elide the assignment, giving us shorter traces. This is particularly useful when the inferred taint goes through an obscure callable, which ends up tainting the whole object.

The tests demonstrate that we now properly shorten issues.

Reviewed By: fahndrich

Differential Revision: D22220100

fbshipit-source-id: 5742dc0c7fd7e78f764a072ed814a478fc5c77a7