Do not throw impossible assertion even when target is None
Summary:
The original effort (D21636197) to remove impossible assertion errors was targeted at allowing users to create "bottomed out" or "impossible" states inside of if-conditions, because those are often useful runtime sanity checks.

Remaining cases of impossible assertion errors were thrown when the target was None and we were checking or asserting that it was not, since None is usually not a given annotation and doesn't fall into the pattern of checking that, say, a parameter is actually typed as the annotation says it is.

However, allowing for sanity checks like this are still valuable if other methods are typed as returning None but may be mistyped, and runtime handling is added in case some value is not None.

Another consideration could be made for impossible calls to `assert` at the toplevel, which would be reasonable to error against if they created a type contradiction because that likely means something was mistyped. However, we can't reasonably distinguish these from an if-statement that intentionally creates a contradictory nested state in which to handle unexpected behavior / cases that were expected to be impossible.

Reviewed By: pradeep90

Differential Revision: D22346214

fbshipit-source-id: 7c7d960de548fafa6d778b394bf230101b2c8b9b