Porting setup.sh to python
Summary:
Maintaining this gnarly bash script has been a maintenance pain for a long time, and has made this problematic to integrate into python workflows, leading to the further proliferation of bash scripts.
In the spirit of trying to find opportunities to dogfood Pyre, I decided to finally take this on.

For compatibility purposes I haven't changed the interface of this at all, and still have maintained setup.sh as a pass through to setup.py, but we should gradually move usages over to directly call setup.py, especially in python callers which can actually call through the object interface, which is even better.

Reviewed By: grievejia

Differential Revision: D22348722

fbshipit-source-id: 959030184bf99417d2d19a38be58f330ca137c2f