Print pyre query terminal output on new line
Summary:
Pyre query results would always print on the same logging line as `Waiting for server...`, which is a message designed to be cleared by the next sticky (success, warning, error) that is printed.

We write to stdout directly with thngs like pyre error output and pyre query output because we want to dump output into the terminal without going through formatting or prefixing via the logging structure. For pyre errors, we will print a clearing log message (either "No errors" or "Found X errors") that avoids this issue.

Options to clean this up are: (1) Print the query output via the logging structure and change logging of sticky output to turn on line wrap temporarily, (2) add a new logging level called "Clear" that doesn't print anything and just wipes active lines, (3) Print some sticky message before we write to stdout.

3rd was the easiest so just went with that for now.

Reviewed By: MaggieMoss

Differential Revision: D22421176

fbshipit-source-id: 90a14c8d55b95b61f3024061f4977f274a58bdad