Rationalizing configure and environment only interface
Summary:
The existing interface for the setup runner was based around directly translating the shell script as directly as possible.  This diff moves some things around in order to make it a more intuitive interface for the callers that we actually want to promote.

This entails removing the   configure_only and environment_only flags from being properties of the runner to being methods that you can execute.  To me, running `configure` is more intuitive than running a runner defined as `configure_only`.  On that same note, I renamed the `__call__` to `full_setup`, and added options to not run the tests, and to run a dune clean.

Reviewed By: dkgi

Differential Revision: D22406436

fbshipit-source-id: 821e25dd75b61ec0a3f5d13b852da2ea61b4d4e3