Add explanation for pyre query failure to look up types
Summary:
Improve a frustrating interaction with pyre query where we fail to look up types by location in a given file sometimes even if the file exists. A common case that causes this is if we never analyzed that file because a .pyi stub shadows it.  In this case, we can't simply return the types in the .pyi file because this query is for returning types by location.

Let's do the next best thing and try to make this readable by identifying when stub shadowing is the problem and erroring with that message.

Reviewed By: dkgi

Differential Revision: D22424085

fbshipit-source-id: fae54949593cc423fec2204b198c3bdd9b752df2