Do not provide type lookup for wildcard import
Summary:
We were returning nonsensical type annotations for wildcard imports, because these get expanded during environment building.

Rather than spend time calculating types for each expression in the expanded import, each of which overwrites the previous and the last of which arbitrarily shows up, let's check for imports that overlap in location and skip those when looking up types.

Reviewed By: grievejia

Differential Revision: D22428310

fbshipit-source-id: f5fb62dfdec9362849661e67b7b560e5a14d35ed