generalize our handling of `assert (x is y)`
Summary:
We have a special case where we refine `x` to be `None` on `assert(x is None)`.
However, `is` is just a general object equality, and thus can be used to refine things to other things than None.
Most commonly it's used to refine metatypes, as a way to do `assert(isinstance(x, Type[Y]))`.

Reviewed By: grievejia

Differential Revision: D22463561

fbshipit-source-id: 6ba19068898fd6d31fbe60538be74bb2b397dba0