expression_contains_any should error about Any in nested types.
Summary:
Problem: We were substituting primitives like `List` and `Callable` at the top-level, but not recursively.

The default visitor returns all names including the generic types, which led to false positives on `Union[Tuple[int]]` since it would find `Tuple` and substitute it with `Tuple[..., Any]`. So wrote a visitor to get the non-generic types.

Reviewed By: dkgi

Differential Revision: D22471140

fbshipit-source-id: cd2b91ce36c0c3bf6b9dfe711fde980418a0ac85