Clarify error messages around generated pyre errors
Summary:
Sometimes the unexpected error output comes from a pyre run that was not piped in by the user. The resulting error message that command "received" invalid JSON as "input" is really confusing/misleading, as is the suggestion to change what gets piped in.

Having a hard time debugging one of these cases with the terse error message here, so clarifying what we print based on where the unexpected string came from, and printing the bad string in the case that it comes from internally generated errors because that's not something the person running this can directly control or have insight into.

Reviewed By: pradeep90

Differential Revision: D22514863

fbshipit-source-id: f6e85fbe28e0a00aea12e7d63d77cc9abdfde094