introduce the `-> SkipOverrides` option for models
Summary:
Context: We currently have a bug with our override registration that only allows registering overrides from a single file (if callables from different files override the same base callable, one file wins, and we only register the first file we see). Fixing this bug revealed significant performance problems in production due to things like `object.__init__` having a large number of overrides.

The "principled" solution to this is to allow users to manually mark a few functions as having unusable override information, so that we can suppress domain-specific use cases in addition to the common object, typing.NamedTuple and typing.Mapping ones. Towards that end, let's introduce this annotation. A future change will fix the bug.

The diff itself is a bit roundabout/complex compared to what it's achieving because of a fundamental reason: The analysis infrastructure has the override registration/information be analysis-agnostic, but all configurations, etc. are per-analysis. So, to support SkipOverrides in taint models, we needed to get more information out of `Analysis.initialize`.

Reviewed By: fahndrich

Differential Revision: D22558432

fbshipit-source-id: 4a8faee2f2db40b8676cd1bda39cbcdc05e2352f