don't register override loops for type
Summary:
We use Pyre's attribute finding mechanism to register overrides. This works by the following mechanism:

When we compute overrides(C), we compute MRO(C), and go through those in order, attempting to find attribute. This works well in most cases, but has an interesting detail.

If we fail to find an attribute for C, we search for the attribute Metaclass(C), as the metaclass might have class attributes which you're looking for, and will have the appropriate parent. However, for `type`, `type(type) = type`, so if we fail to find a method for `type`'s parents (`typing.Generic` and `object`), we'll search for the method in `Metaclass(object)`, which is type, and hit a cycle.

Normally, the metaclass hierarchy and normal hierarchy don't mix, but the fact that `type(type) = type` causes this issue.

Reviewed By: fahndrich

Differential Revision: D22606090

fbshipit-source-id: a12c7da25168f8aad988cf26bfe06caf5e710b6f