Fix pyre upgrade error removal and addition on same line
Summary:
We were not handling cases where an existing fixme (tacked to the end of a line of code rather than put on the line above) needed to be removed and one needed to be added on the same line.

Context:

D22671543:

```
                     # pyre-fixme[6]: `+` is not supported for operand types
                    #  `Optional[str]` and `str`.
                    # pyre-fixme[0]: The `pyre-ignore[16]` or `pyre-fixme[16]`
                    #  comment is not suppressing type errors, please remove it.
                    old_fields.append(col.expr + ' AS ' + col.name)  # pyre-ignore[16]
```

Two lint errors:
Unused ignore [0]: The pyre-ignore[0] or pyre-fixme[0] comment is not suppressing type errors, please remove it.
Unused ignore [0]: The pyre-ignore[16] or pyre-fixme[16] comment is not suppressing type errors, please remove it.

Reviewed By: grievejia

Differential Revision: D22779692

fbshipit-source-id: 0e42d051f2e7a7eee414bee4d64344f4a281a985