Support for Addition and Multiplication on Type Variables bounded by int
Summary:
# Support for Addition and Multiplication on Type Variables bounded by int

This diff adds support for addition and multiplication operations on type variables bounded to int, allowing to type functions such as:

```
def push(v : Vector[N]) -> Vector[Add[N,Literal[1]]]
def duplicate(v : N) -> Multiply[N,Literal[2]]
```

In addition, it introduces an equational theory for simplifying arithmetic expressions on type. When all type variables of the expressions are instantiated with literals it simplifies to the corresponding result:

```
def push(v : Vector[N]) -> Vector[Add[N,Literal[1]]]
v_5 : Vector[Literal[5]]
reveal_type(push(v_5)) # Vector[Literal[6]]
```

On the other hand, when one of the parameters of the arithmetic expression is not a literal but still is bounded by int, then the following rules apply with the following priority.
 1. Add [Any, _] = Any
 2. Add [int, _] = int

# Operators
This diff introduces 2 operators, Add and Multiply, both following the same syntax.

```
Add[A_1, A_2 (,A_i)*]
Multiply[A_1, A_2 (,A_i)*]
#  A_ corresponds to a type variable bounded to int.
```

Any instance of Add with less than 2 type parameters or with parameters not bounded by int will be rejected.

# Equality
Due to the nature of expressions with addition and multiplication, there are infinite ways to express the same function. We recall that addition and multiplication are commutative, distributive, associative and have an identity element.
Nonetheless, this diff guarantees that any 2 expressions that are equivalent will be considered equal from a type perspective, thus accepting cases like the following:

```
def duplicate_with_multiplyt(n : N) -> Multiply[N,Literal[2]] : ...
def duplicate_with_add(n : N) -> Add[N,N] :
    return duplicate_with_multiply(n)
```

To achieve this, aritmethic expressions are rewritten as a list of monomials in graded lexicographic order.

# Limitations
Although equality is totally supported, this approach has some limitations with respect to recognizing subtyping between arithmetic expressions. In particular, the only cases where subtyping can be resolved are the following:

```
# Nums are int literals
# A and B are type variables bounded to int

num1 <: num2
A + num1 <: num2
num1 <: num2 + B
A + num1 <: B + num2
```

In the case of Multiply the rules are the same, but if num2 is not divisible by num1 then there cannot be a subtyping relationship.

And example of a code that would be affected by such limitation is the following:

```
def f(x : Add[N,M], y : M) -> N: ...
f(5,2) # Cannot verify that Literal[5] <: N + M
```

# Implementation
In order to support the exposed changes in Pyre, a new Type has been declared in Type.t, IntExpression of t Polynomial.t.

To represent arbitrary arithmetic expressions on type variables and literals, new modules Polynomial and Monomial have been created. In particular, expressions are represented as a sum of monomials sorted in graded lexicographic order.

All operations between polynomials guarantee that a canonical form is preserved, enabling equality checks. The operations between polynomials that are supported are:
- Addition
- Subtraction
- Multiplication
- Pow
- Replace variable on polynomial

Addition and multiplication are obviously useful to support Add and Multiply. Subtraction is used when needing to resolve subtyping questions such as:

```
A + num1 <: B + num2 -> A + num1 - num1 <: B + num2 - num1
```

Replace variable on polynomial is needed to support local_replace on IntExpressions. Moreover, it relies on pow when x should be replaced in an expression such as x^3 (coming from x * x * x). Note that in that sense pow is just syntax sugar since it only supports positive integer exponents.

# Future extensions

## Preserve literals values on arithmetic
The presented change, together with the idea of supporting any subtype of int would provide all the tools required for enhancing typing of arithmetic expressions of literal values.

```
a : Literal[2] = 2
b : Literal[3] = 3
x = a + b # Currently x : int, proposed x : Literal[5]
```

For that it would be enough to change the type of int._add_ in the following way:

```
def __add__(self : A, other : B) -> Add[A,B]

```
However, there is a concrete case where this approach could be problematic:

```
a : int
b : Any
a.__add__(b)
# Current: int, After change: Add[int,Any] = Any
```

However, it is worth noticing that this issue only arises when __add__ is explicitely typed, otherwise due to precedence the __add__ or __radd__ of Any will be chosen:

```
a : int
b : Any
c = a + b

# Return int
a.__add__(b)

# Return Any
b.__add__(a)
a + b
b + a
c
```

## Fixed Bugs
This diff solves an existing bug related with subtyping of literals integers and enums. An example where this bug occurs is:
https://www.internalfb.com/intern/diffusion/IGSRVHG/browse/master/distillery/status/api/views.py?lines=401-412

Previously, the lower bound of a Literal with any other type that is no the same literal resulted in Bottom. However, now we check if by chance even if they are not the same literal if they have a lower bound different than Bottom. And in the case of Literals of Int and Literals of Enum this turns out to be useful.

Reviewed By: mrkmndz

Differential Revision: D22358132

fbshipit-source-id: 045e983e2f043ee00b9f90a757f0f79900051c6d